pyproject.toml
[tool.poetry]
name = "arcworld"
version = "0.1.0"
description = "Synthetic task generation for the ARC Challange"
authors = ["yassinetb <ybenchekroun@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.26.4"
scipy = "^1.9.3"
pandas = "^2.0.2"
tqdm = "^4.65"
matplotlib = "*"
h5py = "^3.11.0"
factory-boy = "^3.3.0"
pydantic = "^2.12.5"
pillow = "*"

[tool.poetry.group.dev.dependencies]
isort = "*"
ruff = "*"
black = "*"
mypy = "*"

[tool.poetry.group.test.dependencies]
pytest = "*"
#pytest-cov = "*"
torch = "*"

[tool.black]
line-length = 88


[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]